// src/shared/generated/RuntimeRestoreWindowRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RuntimeRestoreTabRecord } from "./RuntimeRestoreTabRecord";
import type { StateWorkspaceDisplayTargetRecord } from "./StateWorkspaceDisplayTargetRecord";

export type RuntimeRestoreWindowRecord = { id: string, targetDisplay: StateWorkspaceDisplayTargetRecord, wasVisible: boolean, activeSourceId?: string, tabs: Array<RuntimeRestoreTabRecord>, };